requirements.txt
numpy
pandas
pyfiglet
PyExifTool
python-yara
